put css files into a separate folder
