+++ front-end/src/App.js
@@ -1,4 +1,4 @@
-import './App.css'
+import './css/App.css'
 import React, { useState, useEffect } from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import Home from './Home.js'
