+++ front-end/src/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Footer.css'
+import './css/Footer.css'
 import { Link } from 'react-router-dom'
 import trakt from './Images/trakt.png'
