+++ front-end/src/HeaderLogo.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import logo from './Images/logo.png'
-import './HeaderLogo.css'
+import './css/HeaderLogo.css'
 
 function HeaderLogo() {
   return (
