+++ front-end/src/Images/palette.png
