+++ front-end/src/LoginButton.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import './LoginButton.css'
+import './css/LoginButton.css'
 
 function LoginButton() {
   return (
