+++ front-end/src/MyShows.js
@@ -4,7 +4,7 @@ import Header from './Header'
 import Footer from './Footer'
 import axios from 'axios'
 import { mockShowAPI, mockShowImage } from './MockData'
-import './MyShows.css'
+import './css/MyShows.css'
 import Modal from 'react-modal'
 import Select from 'react-select'
 import { Link, useHistory } from 'react-router-dom'
