+++ front-end/src/TermsOfService.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Header from './Header'
 import Footer from './Footer'
-import './TermsOfService.css'
+import './css/TermsOfService.css'
 import { Link } from 'react-router-dom'
 
 function TermsOfService() {
