+++ front-end/src/css/IndividualShow.css
